--- a/src/Columns/ColumnCompressed.h
+++ b/src/Columns/ColumnCompressed.h
@@ -107,8 +107,10 @@ public:
     {
         throwMustBeDecompressed();
     }
-    void getPermutation(bool, size_t, int, Permutation &) const override { throwMustBeDecompressed(); }
-    void updatePermutation(bool, size_t, int, Permutation &, EqualRanges &) const override { throwMustBeDecompressed(); }
+    void getPermutation(IColumn::PermutationSortDirection, IColumn::PermutationSortStability,
+                        size_t, int, Permutation &) const override { throwMustBeDecompressed(); }
+    void updatePermutation(IColumn::PermutationSortDirection, IColumn::PermutationSortStability,
+                        size_t, int, Permutation &, EqualRanges &) const override { throwMustBeDecompressed(); }
     ColumnPtr replicate(const Offsets &) const override { throwMustBeDecompressed(); }
     MutableColumns scatter(ColumnIndex, const Selector &) const override { throwMustBeDecompressed(); }
     void gather(ColumnGathererStream &) override { throwMustBeDecompressed(); }