… src/Columns/ColumnDecimal.cpp
… src/Columns/ColumnDecimal.h
… src/Columns/ColumnFixedString.h
… src/Columns/ColumnFunction.h
… src/Columns/ColumnLowCardinality.cpp
… src/Columns/ColumnLowCardinality.h
… src/Columns/ColumnMap.cpp
… src/Columns/ColumnMap.h
… src/Columns/ColumnNullable.cpp
… src/Columns/ColumnNullable.h
… src/Columns/ColumnSparse.cpp
… src/Columns/ColumnSparse.h
… src/Columns/ColumnString.cpp
… src/Columns/ColumnString.h
… src/Columns/ColumnTuple.cpp
… src/Columns/ColumnTuple.h
… src/Columns/ColumnUnique.h
… src/Columns/ColumnVector.cpp
… src/Columns/ColumnVector.h
… src/Columns/IColumn.h
… src/Columns/IColumnDummy.h
… src/Columns/IColumnImpl.h
… src/Columns/tests/gtest_column_sparse.cpp
… src/Columns/tests/gtest_column_stable_permutation.cpp
… src/Interpreters/sortBlock.cpp
… tests/performance/merge_tree_insert.xml