Changelog category (leave one):
- Performance Improvement


Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
MergeTree improve insert performance replacing std::stable_sort with pdqsort.