@kitaisreal 2022-02-22
```sql
CREATE TABLE test_table (id UInt64, value UInt64) ENGINE=MergeTree ORDER BY id;
INSERT INTO test_table SELECT rand64(0), rand64(1) FROM system.numbers LIMIT 1500000;
```
Before optimization
```
0 rows in set. Elapsed: 4.522 sec. Processed 15.73 million rows, 125.83 MB (3.48 million rows/s., 27.82 MB/s.)
```

After optimization:
```
0 rows in set. Elapsed: 3.748 sec. Processed 15.73 million rows, 125.83 MB (4.20 million rows/s., 33.57 MB/s.)
```