@kitaisreal 2022-02-22
```sql
CREATE TABLE test_table (id UInt64, value1 UInt64, value2 UInt64, value3 UInt64) ENGINE=MergeTree ORDER BY id;
INSERT INTO test_table SELECT rand64(0), rand64(1), rand64(2), rand64(3) FROM system.numbers LIMIT 15000000;
```
Before optimization:
```
0 rows in set. Elapsed: 5.056 sec. Processed 15.73 million rows, 125.83 MB (3.11 million rows/s., 24.89 MB/s.)
```
After optimization:
```
0 rows in set. Elapsed: 4.269 sec. Processed 15.73 million rows, 125.83 MB (3.68 million rows/s., 29.47 MB/s.)
```