@kitaisreal 2022-03-13
Final performance results https://s3.amazonaws.com/clickhouse-test-reports/34750/b67f756a438b6896f65f1ab43f174901ccbc4f2c/performance_comparison__actions__[1/4]/report.html.
During insertion in merge tree, if primary key is integer columns, and this is most of the time and whole number of columns in merge tree are not so big, we win around 2-3 times in insertion speed for random data. For not random data pdqsort will work even better.
If there are a lot columns in table, we win around 15-20 percents because most of the time we will spend on columns permutation.
Also interesting case is string primary key, in case of it we spend most of the time on string comparison.